requirements.txt
vosk
sounddevice
pynput
pystray
Pillow
